+++ Source/WebCore/animation/AcceleratedEffectStackUpdater.cpp
@@ -29,7 +29,10 @@
 #if ENABLE(THREADED_ANIMATIONS)
 
 #include "AcceleratedTimeline.h"
+#include "AcceleratedTimelinesUpdater.h"
+#include "DocumentPage.h"
 #include "KeyframeEffectStack.h"
+#include "NodeDocument.h"
 #include "RenderElement.h"
 #include "RenderLayer.h"
 #include "RenderLayerBacking.h"
@@ -45,6 +48,7 @@ void AcceleratedEffectStackUpdater::update()
     if (!hasTargetsPendingUpdate())
         return;
 
+    RefPtr<Page> page;
     HashSet<Ref<AcceleratedTimeline>> timelinesInUpdate;
 
     auto targetsPendingUpdate = std::exchange(m_targetsPendingUpdate, { });
@@ -53,6 +57,9 @@ void AcceleratedEffectStackUpdater::update()
             continue;
 
         Styleable target { *element, pseudoElementIdentifier };
+
+        if (!page)
+            page = element->protectedDocument()->page();
 
         auto* renderer = dynamicDowncast<RenderLayerModelObject>(target.renderer());
         if (!renderer || !renderer->isComposited())
@@ -63,52 +70,13 @@ void AcceleratedEffectStackUpdater::update()
         renderLayer->backing()->updateAcceleratedEffectsAndBaseValues(timelinesInUpdate);
     }
 
-    for (auto& timeline : timelinesInUpdate) {
-        auto& timelineIdentifier = timeline->identifier();
-        auto addResult = m_timelines.add(timelineIdentifier, timeline.ptr());
-        if (addResult.isNewEntry)
-            m_timelinesUpdate.created.add(timeline);
-    }
+    ASSERT(page);
+    page->ensureAcceleratedTimelinesUpdater().processTimelinesSeenDuringEffectStacksUpdate(WTFMove(timelinesInUpdate));
 }
 
 void AcceleratedEffectStackUpdater::scheduleUpdateForTarget(const Styleable& target)
 {
     m_targetsPendingUpdate.add({ &target.element, target.pseudoElementIdentifier });
-}
-
-void AcceleratedEffectStackUpdater::scrollTimelineDidChange(ScrollTimeline& timeline)
-{
-    m_scrollTimelinesPendingUpdate.add(timeline);
-}
-
-AcceleratedTimelinesUpdate AcceleratedEffectStackUpdater::takeTimelinesUpdate()
-{
-    // All known accelerated timelines that got destroyed since the last update
-    // will now be null references. Add them to the list of destroyed timelines.
-    for (auto& [timelineIdentifier, timeline] : m_timelines) {
-        if (!timeline)
-            m_timelinesUpdate.destroyed.add(timelineIdentifier);
-    }
-
-    // Prune all those destroyed timelines from our list of know accelerated timelines.
-    for (auto& identifierToRemove : m_timelinesUpdate.destroyed)
-        m_timelines.remove(identifierToRemove);
-
-    // Finally, process all timelines that were marked as requiring an update, either
-    // marking them as modified or destroyed if they no longer are accelerated.
-    auto scrollTimelinesPendingUpdate = std::exchange(m_scrollTimelinesPendingUpdate, { });
-    for (auto& scrollTimeline : scrollTimelinesPendingUpdate) {
-        auto timelineIdentifier = scrollTimeline->acceleratedTimelineIdentifier();
-        auto acceleratedTimeline = m_timelines.getOptional(timelineIdentifier);
-        if (acceleratedTimeline && scrollTimeline->canBeAccelerated()) {
-            ASSERT(*acceleratedTimeline);
-            scrollTimeline->updateAcceleratedRepresentation();
-            m_timelinesUpdate.modified.add(**acceleratedTimeline);
-        } else
-            m_timelinesUpdate.destroyed.add(timelineIdentifier);
-    }
-
-    return std::exchange(m_timelinesUpdate, { });
 }
 
 } // namespace WebCore
