+++ Source/WebCore/animation/AcceleratedTimelinesUpdater.h
@@ -1,0 +1,56 @@
+/*
+ * Copyright (C) 2025 Apple Inc. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS''
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+ * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS
+ * BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
+ * THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#pragma once
+
+#if ENABLE(THREADED_ANIMATIONS)
+
+#include <WebCore/AcceleratedTimeline.h>
+#include <WebCore/AnimationMalloc.h>
+#include <WebCore/TimelineIdentifier.h>
+#include <wtf/HashSet.h>
+
+namespace WebCore {
+
+class ScrollTimeline;
+
+class AcceleratedTimelinesUpdater {
+    WTF_DEPRECATED_MAKE_FAST_ALLOCATED_WITH_HEAP_IDENTIFIER(AcceleratedTimelinesUpdater, Animation);
+public:
+    AcceleratedTimelinesUpdater() = default;
+
+    void scrollTimelineDidChange(ScrollTimeline&);
+    void processTimelinesSeenDuringEffectStacksUpdate(HashSet<Ref<AcceleratedTimeline>>&&);
+    WEBCORE_EXPORT AcceleratedTimelinesUpdate takeTimelinesUpdate();
+
+private:
+    HashSet<Ref<ScrollTimeline>> m_scrollTimelinesPendingUpdate;
+    HashMap<TimelineIdentifier, WeakPtr<AcceleratedTimeline>> m_timelines;
+    AcceleratedTimelinesUpdate m_timelinesUpdate;
+};
+
+} // namespace WebCore
+
+#endif // ENABLE(THREADED_ANIMATIONS)
