+++ Source/WebCore/animation/ScrollTimeline.cpp
@@ -419,9 +419,11 @@ bool ScrollTimeline::computeCanBeAccelerated() const
 
 void ScrollTimeline::scheduleAcceleratedRepresentationUpdate()
 {
-    if (CheckedPtr controller = this->controller()) {
-        if (auto* acceleratedEffectStackUpdater = controller->existingAcceleratedEffectStackUpdater())
-            acceleratedEffectStackUpdater->scrollTimelineDidChange(*this);
+    if (auto source = m_source.styleable()) {
+        if (RefPtr page = source->element.protectedDocument()->page()) {
+            if (auto* acceleratedTimelinesUpdater = page->acceleratedTimelinesUpdater())
+                acceleratedTimelinesUpdater->scrollTimelineDidChange(*this);
+        }
     }
 }
